Separate module for api entities.
