+++ lib/mongo_browser.rb
@@ -29,3 +29,4 @@ require "mongo_browser/models/document"
 
 require "mongo_browser/application"
 require "mongo_browser/api"
+require "mongo_browser/api/entities"
